% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{az_account}
\alias{az_account}
\alias{az_account_list}
\alias{az_account_show}
\alias{az_account_set}
\title{Manage Azure subscription information}
\usage{
az_account(subscription = NULL, ...)

az_account_list(...)

az_account_show(subscription = NULL, ...)

az_account_set(subscription, ...)
}
\arguments{
\item{subscription}{Name or ID of the Azure subscription to which costs are billed.
According to an upvoted answer on Stack Overflow, \href{https://stackoverflow.com/questions/45661109/are-azure-subscription-id-aad-tenant-id-and-aad-app-client-id-considered-secre}{Azure subscription IDs need not be considered a secret or personal identifiable information (PII)}.
However, depending your applicable context and policies, you may want to provide this argument as a secret.

To find out which subscriptions you are currently authorised to use, run \code{print(az_account_list())}.}

\item{...}{
  Arguments passed on to \code{\link[=az_cli_run]{az_cli_run}}
  \describe{
    \item{\code{cmd}}{Command, required, optional, additional, extra parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments to the Azure CLI.
\code{extra} parameters are reserved for internal use.}
    \item{\code{req}}{Command, required, optional, additional, extra parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments to the Azure CLI.
\code{extra} parameters are reserved for internal use.}
    \item{\code{opt}}{Command, required, optional, additional, extra parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments to the Azure CLI.
\code{extra} parameters are reserved for internal use.}
    \item{\code{add}}{Command, required, optional, additional, extra parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments to the Azure CLI.
\code{extra} parameters are reserved for internal use.}
    \item{\code{echo_cmd}}{Whether to print the command to run to the screen.}
    \item{\code{echo}}{Whether to print the standard output and error
to the screen. Note that the order of the standard output and error
lines are not necessarily correct, as standard output is typically
buffered.}
  }}
}
\description{
Subscription defaults to \code{NULL}, in which case the subscription is expected to be enabled in the azure CLI cache already.
Appends and sets to default \code{subscription}, if provided.
Errors out if no subscription is enabled.
}
\details{
Subscriptions are kept in the local azure CLI cache, so you should not have to run this more than once.
On GitHub Actions, \href{https://github.com/azure/login}{the azure login action} will already set up a subscription.
}
\section{Functions}{
\itemize{
\item \code{az_account_list}: Get a list of subscriptions for the logged in account.

\item \code{az_account_show}: Get the details of a subscription.
If the subscription isn't specified, shows the details of the default subscription.

\item \code{az_account_set}: Set a subscription to be the current active subscription.
}}

