% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.r
\name{az_login}
\alias{az_login}
\title{Log in to Azure}
\usage{
az_login(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=az_cli_run]{az_cli_run}}
  \describe{
    \item{\code{cmd}}{Command, required, optional, additional, extra parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments to the Azure CLI.
\code{extra} parameters are reserved for internal use.}
    \item{\code{req}}{Command, required, optional, additional, extra parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments to the Azure CLI.
\code{extra} parameters are reserved for internal use.}
    \item{\code{opt}}{Command, required, optional, additional, extra parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments to the Azure CLI.
\code{extra} parameters are reserved for internal use.}
    \item{\code{add}}{Command, required, optional, additional, extra parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments to the Azure CLI.
\code{extra} parameters are reserved for internal use.}
    \item{\code{echo_cmd}}{Whether to print the command to run to the screen.}
    \item{\code{echo}}{Whether to print the standard output and error
to the screen. Note that the order of the standard output and error
lines are not necessarily correct, as standard output is typically
buffered.}
  }}
}
\description{
Helper for \strong{interactive} login.
}
