% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ghactions_url}
\alias{get_ghactions_url}
\title{Get the URL to GitHub Actions for a local repo.}
\usage{
get_ghactions_url(path = ".")
}
\arguments{
\item{path}{Path that is contained within a git repo.}
}
\description{
Get the URL to GitHub Actions for a local repo.
}
\keyword{internal}
